test($CarNameValidationTest): update test method logic
- new CarName() -> nameValidator.validate()
